Add user.Encode() function
